still not connecting

+++ MVCPerson/Models/Services/PeopleService.cs
@@ -1,4 +1,7 @@
-﻿using MVCPerson.Models;
+﻿using System;
+using System.Linq;
+using System.Collections.Generic;
+using System.Threading.Tasks;
 using MVCPerson.Models.ViewModels;
 using MVCPerson.Models.Repos;
 
@@ -6,15 +9,14 @@ namespace MVCPerson.Models.Services
 {
     public class PeopleService : IPeopleService
     {
-        readonly IPeopleRepo _peopleRepo;
+        IPeopleRepo _peopleRepo;
         public PeopleService(IPeopleRepo peopleRepo)
         {
             _peopleRepo = peopleRepo;
         }
 
-        public Person Add(CreatePersonViewModel createPerson)
+        public Person Create(CreatePersonViewModel createPerson)
         {
-           // Person person = _peopleRepo.Create(createPerson.Name, createPerson.PhoneNumber, createPerson.CityName);
 
             if (string.IsNullOrWhiteSpace(createPerson.Name)
                             || string.IsNullOrWhiteSpace(createPerson.PhoneNumber)
@@ -30,24 +32,24 @@ namespace MVCPerson.Models.Services
                 CityName = createPerson.CityName,
 
             };
-            person = _peopleRepo.Create(person);
+           _peopleRepo.Create(person);
             return person;
         }
 
-        public List<Person> All()
+        public List<Person> GetAll()
         {
-            return _peopleRepo.Read();
+            return _peopleRepo.GetAll();
         }
         public Person FindById(int id)
         {
-            Person foundperson = _peopleRepo.Read(id);
+            Person foundperson = _peopleRepo.GetById(id);
             return foundperson;
 
         }
 
         public bool Edit(int id, CreatePersonViewModel editPerson)
         {
-            Person orginalPerson = FindById(id);
+            Person orginalPerson = _peopleRepo.GetById(id);
             if(orginalPerson != null)
             {
                 orginalPerson.Name = editPerson.Name;
@@ -55,22 +57,22 @@ namespace MVCPerson.Models.Services
                 orginalPerson.CityName = editPerson.CityName;
             }
             return _peopleRepo.Update(orginalPerson);
-
         }
 
-        public bool Remove(int id)
+        public void  Remove(int id)
         {
-            Person personToDelete = _peopleRepo.Read(id);
-            bool success = _peopleRepo.Delete(personToDelete);
-
-            return success;
+            Person personToDelete = _peopleRepo.GetById(id);
+            if(personToDelete != null)
+            {
+              _peopleRepo.Delete(personToDelete);
+            }
 
         }
         public List<Person> Search(string search)
         {
-            List<Person> searchPerson = _peopleRepo.Read();
+            List<Person> searchPerson = _peopleRepo.GetAll();
             {
-                foreach (Person item in _peopleRepo.Read())
+                foreach (Person item in _peopleRepo.GetAll())
                 {
                     if (item.Name.Contains(search, StringComparison.OrdinalIgnoreCase)
                       || item.CityName.Contains(search, StringComparison.OrdinalIgnoreCase))
